- Migrate finished non-video torrents to desktop directory.  - Changed downloader interface to take whether or not to delete files    on a stop and what directory to migrate files to.  - Changed My Collection and New Videos filters (portable/filters.py)    to not show non-video torrents.  - Added shortFilename field to FileItem to enable:  - Fixed incorrect migration of directories of videos.  - Made Completed (and erred) non-video Torrents remain in Downloading area.  - Added Clear Torrents button.
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2651 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: e52bb2a4cef630d67d66c44593ef30809378df87

+++ tv/portable/dl_daemon/download.py
@@ -190,7 +190,7 @@ def startDownload(dlid):
         return True
     return download.start()
 
-def stopDownload(dlid):
+def stopDownload(dlid, delete):
     try:
         _lock.acquire()
         try:
@@ -200,16 +200,16 @@ def stopDownload(dlid):
             _lock.release()
     except: # There is no download with this id
         return True
-    return download.stop()
+    return download.stop(delete)
 
-def migrateDownload(dlid):
+def migrateDownload(dlid, directory):
     try:
         download = _downloads[dlid]
     except: # There is no download with this id
         pass
     else:
         if download.state in ("finished", "uploading"):
-            download.moveToMoviesDirectory()
+            download.moveToDirectory(directory)
             download.updateClient()
 
 def getDownloadStatus(dlids = None):
@@ -324,11 +324,14 @@ class BGDownloader:
         """Move our downloaded file from the Incomplete Downloads directoy to
         the movies directory.
         """
-
         if chatter:
             print "moving to movies directory filename is ", self.filename
-        newfilename = os.path.join(config.get(prefs.MOVIES_DIRECTORY),
-                self.shortFilename)
+        self.moveToDirectory(config.get(prefs.MOVIES_DIRECTORY))
+
+    def moveToDirectory (self, directory):
+        newfilename = os.path.join(directory, self.shortFilename)
+        if newfilename == self.filename:
+            return
         newfilename = nextFreeFilename(newfilename)
         try:
             shutil.move(self.filename, newfilename)
@@ -583,7 +586,7 @@ class HTTPDownloader(BGDownloader):
     ##
     # Stops the download and removes the partially downloaded
     # file.
-    def stop(self):
+    def stop(self, delete):
         if self.state == "downloading":
             if self.filehandle is not None:
                 try:
@@ -592,6 +595,14 @@ class HTTPDownloader(BGDownloader):
                     remove(self.filename)
                 except:
                     pass
+        if delete:
+            try:
+                if os.path.isdir(self.filename):
+                    shutil.rmtree(self.filename)
+                else:
+                    remove(self.filename)
+            except:
+                pass
         self.currentSize = 0
         self.cancelRequest()
         self.state = "stopped"
@@ -764,13 +775,13 @@ class BTDownloader(BGDownloader):
                 'fractionDone': fractionDone, 'upTotal': upTotal, 
                 'errorTime': errorTime, 'errorMessage': errorMessage}
 
-    def moveToMoviesDirectory(self):
+    def moveToDirectory(self, directory):
         if self.state in ('uploading', 'downloading'):
             self._shutdownTorrent()
-            BGDownloader.moveToMoviesDirectory(self)
+            BGDownloader.moveToDirectory(self, directory)
             self._startTorrent()
         else:
-            BGDownloader.moveToMoviesDirectory(self)
+            BGDownloader.moveToDirectory(self, directory)
 
     def restoreState(self, data):
         self.__dict__ = data
@@ -804,17 +815,18 @@ class BTDownloader(BGDownloader):
         except AttributeError:
             pass
 
-    def stop(self):
+    def stop(self, delete):
         self.state = "stopped"
         self.updateClient()
         self._shutdownTorrent()
-        try:
-            if os.path.isdir(self.filename):
-                shutil.rmtree(self.filename)
-            else:
-                remove(self.filename)
-        except:
-            pass
+        if delete:
+            try:
+                if os.path.isdir(self.filename):
+                    shutil.rmtree(self.filename)
+                else:
+                    remove(self.filename)
+            except:
+                pass
 
     def start(self):
         self.pause()
